backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
drf-spectacular==0.27.2
gunicorn==21.2.0
whitenoise==6.5.0
